Move position & hashable types to incrementalmerkletree root.
