+++ Cargo.toml
@@ -1,4 +1,5 @@
 [workspace]
 members = [
+    "incrementalmerkletree",
     "bridgetree",
 ]
