+++ incrementalmerkletree/Cargo.toml
@@ -1,24 +1,23 @@
 [package]
-name = "bridgetree"
-version = "0.2.0"
+name = "incrementalmerkletree"
+version = "0.3.0"
 authors = [
-    "Kris Nuttycombe <kris@nutty.land>",
     "Sean Bowe <ewillbefull@gmail.com>",
+    "Kris Nuttycombe <kris@nutty.land>",
 ]
 edition = "2018"
 license = "MIT OR Apache-2.0"
-description = "A space-efficient Merkle tree designed for linear appends with witnessing of marked leaves, checkpointing & state restoration."
+description = "Common types, interfaces, and utilities for Merkle tree data structures"
 homepage = "https://github.com/zcash/incrementalmerkletree"
 repository = "https://github.com/zcash/incrementalmerkletree"
 categories = ["algorithms", "data-structures"]
 
 [dependencies]
-incrementalmerkletree = { version = "0.3", path = "../incrementalmerkletree" }
+either = "1.8"
 serde = { version = "1", features = ["derive"] }
 proptest = { version = "1.0.0", optional = true }
 
 [dev-dependencies]
-incrementalmerkletree = { version = "0.3", path = "../incrementalmerkletree", features = ["test-dependencies"] }
 proptest = "1.0.0"
 
 [features]
