+++ incrementalmerkletree/README.md
@@ -1,0 +1,6 @@
+incrementalmerkletree
+=====================
+
+This crate contains common utility types and testing infrastructure used in the implementation
+of the `bridgetree` and `shardtree` crates.
+
