+++ incrementalmerkletree/src/testing.rs
@@ -1,0 +1,30 @@
+use crate::{Hashable, Level};
+
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub struct SipHashable(pub u64);
+
+impl Hashable for SipHashable {
+    fn empty_leaf() -> Self {
+        SipHashable(0)
+    }
+
+    fn combine(_level: Level, a: &Self, b: &Self) -> Self {
+        #![allow(deprecated)]
+        use std::hash::{Hasher, SipHasher};
+
+        let mut hasher = SipHasher::new();
+        hasher.write_u64(a.0);
+        hasher.write_u64(b.0);
+        SipHashable(hasher.finish())
+    }
+}
+
+impl Hashable for String {
+    fn empty_leaf() -> Self {
+        "_".to_string()
+    }
+
+    fn combine(_: Level, a: &Self, b: &Self) -> Self {
+        a.to_string() + b
+    }
+}
